Remove useless lines caused by declaration-after-statement warning
This is done using the same automated technique introduced for the Citus
repo in https://github.com/citusdata/citus/pull/3181. The main idea is
to replace:

```c
int a = 0;
// some code
a = 1;
```

with:

```c
// some code
int a = 1;
```

For details of how this works see the above mentioned Citus PR.

This comment about not doing early an early return to limit the scope of
variables caused me to add this here too:
https://github.com/citusdata/pg_auto_failover/pull/244#discussion_r424306388
With this commit merged, scope of variables will always be minimal in
cases where an early return is possible.